requirements.txt
requests
pandas
pyyaml
tqdm
openai
seaborn